rename timer_queue module to timer
