+++ xernel/kernel/src/timer/timer_event.rs
@@ -1,9 +1,10 @@
-use crate::{arch::amd64::interrupts::dpc::Dpc, current_cpu};
+use crate::{arch::amd64::interrupts::dpc::Dpc, current_cpu}; 
 use alloc::boxed::Box;
 
 pub trait EventExecutor {
     fn dispatch(self: Box<Self>);
     fn deadline(&self) -> usize;
+    fn periodic(&self) -> bool;
 }
 
 enum EventState {
@@ -12,14 +13,12 @@ enum EventState {
 }
 
 pub struct TimerEvent<T> {
-   // callback: fn(T),
-   // data: T,
     dpc: Dpc<T>,
     //    nanosecs: usize,
     deadline: usize,
     state: EventState,
     callback_core: u32,
-    //    periodic: bool,
+    periodic: bool,
 }
 
 impl<T: 'static> EventExecutor for TimerEvent<T> {
@@ -30,16 +29,21 @@ impl<T: 'static> EventExecutor for TimerEvent<T> {
     fn deadline(&self) -> usize {
         self.deadline
     }
+
+    fn periodic(&self) -> bool {
+        self.periodic
+    }
 }
 
 impl<T> TimerEvent<T> {
-    pub fn new(callback: fn(T), data: T, deadline: usize) -> Self {
+    pub fn new(callback: fn(T), data: T, deadline: usize, periodic: bool) -> Self {
         let dpc = Dpc::new(callback, data);
         Self {
             dpc,
             deadline,
             state: EventState::Waiting,
             callback_core: current_cpu().lapic_id,
+            periodic
         }
     }
 }
