+++ xernel/kernel/src/timer/timer_queue.rs
@@ -3,9 +3,10 @@ use crate::arch::amd64::interrupts::allocate_vector;
 use crate::arch::amd64::interrupts::ipl::IPL;
 use crate::arch::amd64::interrupts::register_handler;
 use crate::cpu::current_cpu;
+use crate::dbg;
 use crate::sched::context::TrapFrame;
-use crate::timer_queue::timer_event::EventExecutor;
-use crate::timer_queue::timer_event::TimerEvent;
+use crate::timer::timer_event::EventExecutor;
+use crate::timer::timer_event::TimerEvent;
 use alloc::boxed::Box;
 use alloc::collections::VecDeque;
 use libxernel::sync::Once;
@@ -31,6 +32,10 @@ impl TimerQueue {
     }
 
     pub fn queue_event<T: 'static>(&mut self, event: TimerEvent<T>) {
+        if self.events.len() == 0 {
+            APIC.oneshot(*TIMER_VECTOR, event.deadline() as u64);
+        }
+
         let insert_index = self
             .events
             .iter()
@@ -53,19 +58,27 @@ pub fn timer_interrupt_handler(frame: &mut TrapFrame) {
     // create dpc and add to queue
     // if periodic, add again to queue
     // set timer to next event in queue
+    //
+    debug!("Handler for timer event queue");
 
     let cpu = current_cpu();
     let mut timer_queue = cpu.timer_queue.write();
 
     timer_queue.event_dispatch();
 
-    let next_event = cpu.timer_queue.read().events.front();
+    let next_event = timer_queue.events.front();
 
     if let Some(event) = next_event {
 
-        APIC.oneshot(*TIMER_VECTOR, event.deadline());
+        APIC.oneshot(*TIMER_VECTOR, event.deadline() as u64);
+
+        if event.periodic() {
+  //          timer_queue.queue_event(event.clone());
+        }
 
     } else {
         // No event in event queue?
     }
+
+    debug!("ENd of timer event queue handler");
 }
